added control for postprocessing, and threaded event processing

+++ postprocessing/eventprocessing.h
@@ -28,7 +28,7 @@ public:
 	EventQueue::simInfo *getDataEvent();
 	void processBinnedEvents(double timeResolution, std::string path, int mapResolution, double zThresshold);
 
-	void processEvent(EventQueue::dataEvent *event, double thresshold, double mapRes, double timeRes, std::string path);
+	void processEvent(EventQueue::dataEvent *event, double thresshold, int mapRes, double timeRes, std::string path);
 private:
 
 	std::vector<EventQueue::dataEvent> eventbin;
@@ -37,7 +37,7 @@ private:
 	EventQueue::simInfo *simInfo;
 	EventQueue::dataEvent devent;
 
-	void recursiveZlevel(AutonLUA *auton, EventQueue::dataEvent *event, std::set<std::string> *visited, int x, int y, int width, int height, double mapRes, double timeRes, double thressholdZ);
+	void recursiveZlevel(AutonLUA *auton, EventQueue::dataEvent *event, std::set<std::string> *visited, int x, int y, int width, int height, int mapRes, double timeRes, double thressholdZ);
 };
 
 #endif // EVENTPROCESSING_H
